+++ tsconfig.json
@@ -9,7 +9,7 @@
   },
   "include": [
     "src/**/*"
-  ],
+, "examples", "sandbox-examples"  ],
   "exclude": [
     "node_modules"
   ]
